requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
openpyxl
requests
pyspellchecker>=0.7.0